커밋

+++ tomatoFarm_server/src/main/java/com/example/demo/user/user_address/entity/UserAddress.java
@@ -16,7 +16,7 @@ public class UserAddress {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long seq;
+    private Integer seq;
     @Column(name = "user_id")
     private String userId;
     private String info;
